explicit archive storage structure
