+++ test/test.vcxproj
@@ -136,10 +136,11 @@
   <ItemGroup>
     <ClInclude Include="..\include\elastic.hpp" />
     <ClInclude Include="..\include\elastic\archive\archive_exception.hpp" />
+    <ClInclude Include="..\include\elastic\archive\basic_oprimitive.hpp" />
     <ClInclude Include="..\include\elastic\archive\basic_primitive.hpp" />
     <ClInclude Include="..\include\elastic\archive\binary_iarchive.hpp" />
     <ClInclude Include="..\include\elastic\archive\binary_oarchive.hpp" />
-    <ClInclude Include="..\include\elastic\archive\binary_primitive.hpp" />
+    <ClInclude Include="..\include\elastic\archive\basic_iprimitive.hpp" />
     <ClInclude Include="..\include\elastic\archive\interface_archive.hpp" />
     <ClInclude Include="..\include\elastic\archive\iserialize.hpp" />
     <ClInclude Include="..\include\elastic\archive\oserialize.hpp" />
