+++ test/test.vcxproj.filters
@@ -110,7 +110,7 @@
     <ClInclude Include="exception.h">
       <Filter>头文件</Filter>
     </ClInclude>
-    <ClInclude Include="..\include\elastic\archive\binary_primitive.hpp">
+    <ClInclude Include="..\include\elastic\archive\basic_iprimitive.hpp">
       <Filter>elastic\archive</Filter>
     </ClInclude>
     <ClInclude Include="..\include\elastic.hpp">
@@ -120,6 +120,9 @@
       <Filter>elastic\archive</Filter>
     </ClInclude>
     <ClInclude Include="..\include\elastic\archive\oserialize.hpp">
+      <Filter>elastic\archive</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\elastic\archive\basic_oprimitive.hpp">
       <Filter>elastic\archive</Filter>
     </ClInclude>
   </ItemGroup>
